Fix tests for StartPage by splitting into two modules: CompanyView and RegionalView
